+++ qbraid/runtime/native/result.py
@@ -17,10 +17,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Optional
 
+from qbraid.exceptions import QbraidError
 from qbraid.runtime.result import AnnealingResultData, GateModelResultData
 
 if TYPE_CHECKING:
     import flair_visual.animation.runtime.qpustate
+    import flair_visual.simulation_result
     import pandas as pd
 
 
@@ -28,46 +30,58 @@ class QuEraQasmSimulatorResultData(GateModelResultData):
     """Class for storing and accessing the results of a QuEra QASM simulator job."""
 
     def __init__(
-        self,
-        backend: str,
-        flair_visual_version: Optional[str] = None,
-        atom_animation_state: Optional[dict[str, Any]] = None,
-        logs: Optional[list[dict[str, Any]]] = None,
-        **kwargs,
+        self, backend: str, quera_simulation_result: Optional[dict[str, Any]] = None, **kwargs
     ):
         """Create a new QuEraSimulatorResultData instance."""
         super().__init__(**kwargs)
         self._backend = backend
-        self._flair_visual_version = flair_visual_version
-        self._atom_animation_state = atom_animation_state
-        self._logs = logs if logs else []
+        self._quera_simulation_result: Optional[
+            flair_visual.simulation_result.QuEraSimulationResult
+        ] = (
+            self._initialize_simulation_result(quera_simulation_result)
+            if quera_simulation_result
+            else None
+        )
 
     @property
     def backend(self) -> str:
         """Returns the backend used."""
         return self._backend
 
+    def _initialize_simulation_result(
+        self, value: dict[str, Any]
+    ) -> flair_visual.simulation_result.QuEraSimulationResult:
+        """Initialize the QuEra simulation result during construction."""
+        try:
+            # pylint: disable-next=import-outside-toplevel
+            from flair_visual.simulation_result import QuEraSimulationResult
+
+            return QuEraSimulationResult.from_json(value)
+        except ImportError as err:  # pragma: no cover
+            raise QbraidError(
+                "Missing required package 'flair-visual'. "
+                "Install with 'pip install flair-visual' or 'pip install qbraid[quera]'."
+            ) from err
+
     @property
-    def flair_visual_version(self) -> Optional[str]:
+    def quera_simulation_result(self) -> flair_visual.simulation_result.QuEraSimulationResult:
+        """Returns the QuEra simulation result."""
+        if self._quera_simulation_result is None:
+            raise ValueError("The simulation result is not available.")
+        return self._quera_simulation_result
+
+    @property
+    def flair_visual_version(self) -> str:
         """Returns the version of the flair visualizer used."""
-        return self._flair_visual_version
+        return self.quera_simulation_result.flair_visual_version
 
     def get_qpu_state(self) -> flair_visual.animation.runtime.qpustate.AnimateQPUState:
-        """Returns the the state of the QPU atoms used in the simulation."""
-        # pylint: disable-next=import-outside-toplevel
-        from flair_visual.animation.runtime.qpustate import AnimateQPUState
-
-        if self._atom_animation_state is None:
-            raise ValueError("No atom_animation_state found in the result data.")
-
-        return AnimateQPUState.from_json(self._atom_animation_state)
+        """Returns the state of the QPU atoms used in the simulation."""
+        return self.quera_simulation_result.atom_animation_state
 
     def get_logs(self) -> pd.DataFrame:
         """Returns the logs generated during the simulation as a pandas DataFrame."""
-        # pylint: disable-next=import-outside-toplevel
-        from pandas import DataFrame
-
-        return DataFrame(self._logs)
+        return self.quera_simulation_result.logs
 
 
 class QbraidQirSimulatorResultData(GateModelResultData):